feat: reorganize codebase into modular structure, add asset and game state management, implement mobile controls and settings menu
