+++ src/core/assets.js
